+++ nevergrad/parametrization/test_parameter.py
@@ -403,3 +403,12 @@ def test_array_sampling(method: str, exponent: tp.Optional[float], sigma: float)
         assert np.any(np.abs(val) > 10)
         assert np.all(val <= mbound)
         assert np.all(val >= 1)
+
+
+def test_parenthood() -> None:
+    param = par.Instrumentation(par.Scalar(init=1.0, mutable_sigma=True).set_mutation(exponent=2.0, sigma=5))
+    sigma_uid = param[0][0].sigma.uid  # type: ignore
+    param_samp = param.sample()
+    param_spawn = param.spawn_child()
+    assert param_samp[0][0].sigma.parents_uids == []  # type: ignore
+    assert param_spawn[0][0].sigma.parents_uids == [sigma_uid]  # type: ignore